IDVA6-368 Unit tests refactored to separate them to their own files for making rebasing easier
